backend/requirements.txt
# AdhesioSense AI Server Dependencies
# Python 3.8+ required

# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Image Processing
Pillow==10.0.1
numpy==1.24.3

# Machine Learning/AI (Optional - for actual model integration)
# tensorflow==2.13.0
# keras==2.13.1
# opencv-python==4.8.1.78
# scikit-learn==1.3.0

# Development/Testing
gunicorn==21.2.0
python-dotenv==1.0.0

# Note: For production deployment with actual AI models, uncomment the ML dependencies
# and install appropriate version of TensorFlow/TensorFlow Lite based on your hardware